requirements.txt
python==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
huggingface-hub==0.12.0
matplotlib==3.5.2
numpy==1.23.3
opencv-python==4.6.0.66
pandas==1.4.4
Pillow==9.3.0
tokenizers==0.13.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
gensim==4.3.2
wordcloud==1.9.3
scipy==1.11.1
diffusers==0.25.1
keybert==0.7.0


